Feat : 2진계산기 9자리 구현
- 9자리 구현하기 위한 메서드 추가
- UInt타입으로 계산 수정

+++ Calculator/Calculator/CustomPrecedenceGroup.swift
@@ -11,11 +11,11 @@ import Foundation
 infix operator ~& : MultiplicationPrecedence
 infix operator ~| : AdditionPrecedence
 
-func ~& (lhs: UInt8, rhs: UInt8) -> UInt8 {
+func ~& (lhs: UInt, rhs: UInt) -> UInt {
     return ~lhs | ~rhs
 }
 
-func ~| (lhs: UInt8, rhs: UInt8) -> UInt8 {
+func ~| (lhs: UInt, rhs: UInt) -> UInt {
     return ~lhs & ~rhs
 }
